move builder trait into builder module
